Only copy files if different

+++ bash/dotfiles/update-dotfiles.sh
@@ -4,6 +4,8 @@ SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]:-$0}"; )" &> /dev/null && 
 declare -a DOTFILES=("$HOME/.bash_profile" "$HOME/.bashrc" "$HOME/.emacs" "$HOME/.vimrc" "$HOME/.tmux.conf")
 
 for dotfile in "${DOTFILES[@]}"; do
-	cp -vi "$dotfile" "$SCRIPT_DIR"
+	if ! cmp -s "$dotfile" "$SCRIPT_DIR/$(basename "$dotfile")"; then
+		cp -vi "$dotfile" "$SCRIPT_DIR"
+	fi
 done
